requirements.txt
Django==4.2.7
djangorestframework==3.14.0
drf-yasg==1.21.7
gunicorn==21.2.0
pillow==10.1.0
python-decouple==3.8